Updated factor_graph.py
